% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_outputStructureFunctions.R
\name{allVarMatchMessage}
\alias{allVarMatchMessage}
\title{allVarMatchMessage}
\usage{
allVarMatchMessage(x)
}
\arguments{
\item{x}{An dataCompareR object}
}
\value{
A string containing the required message
}
\description{
Returns data about matching
}

