% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_checkEmpty.R
\name{checkEmpty}
\alias{checkEmpty}
\title{checkEmpty}
\usage{
checkEmpty(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
None. Stops if empty.
}
\description{
checkEmpty
}
\examples{
\dontrun{checkEmpty(iris)}
}
