% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_checkColumnNames.R
\name{checkForRCompareCol}
\alias{checkForRCompareCol}
\title{checkForRcompareCol}
\usage{
checkForRCompareCol(df1)
}
\arguments{
\item{df1}{a data frame}
}
\value{
None. Stops if error.
}
\description{
checkForRcompareCol
}
\examples{
\dontrun{checkForRcompareCol(iris)}
}
