% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_checkKeysExists.R
\name{checkKeysExist}
\alias{checkKeysExist}
\title{checkKeysExist}
\usage{
checkKeysExist(df, keys)
}
\arguments{
\item{df}{a data frame}

\item{keys}{a list of expected columns}
}
\value{
None. Stops if keys are not present as column names in df.
}
\description{
checkKeysExist
}
\examples{
\dontrun{checkKeysExist(iris, 'columnName')}
}

