% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_checkEmpty.R
\name{checkNA}
\alias{checkNA}
\title{CheckNA}
\usage{
checkNA(df)
}
\arguments{
\item{df}{A (probable) dataframe}
}
\value{
Nothing. Errors is df is NA
}
\description{
Checks a data frame is NA - if so, stops
}
