% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_matchRows.R
\name{checkUniqueness}
\alias{checkUniqueness}
\title{Checks that a list of indexes areunique}
\usage{
checkUniqueness(df_indices)
}
\arguments{
\item{df_indices}{A vector of values}
}
\value{
Boolean - true if all values in vector are unique, false if not
}
\description{
Checks that a list of indexes areunique
}
\examples{
\dontrun{checkUniqueness(c('car','van','van'))}
\dontrun{checkUniqueness(c('car','van','bus'))}
}

