% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_matchColumns.R
\name{cleanColNames}
\alias{cleanColNames}
\title{cleanColNames : get colnames, remove leading and trailing whitespace and push to upper case}
\usage{
cleanColNames(DF)
}
\arguments{
\item{DF}{Input dataframe}
}
\value{
colInfo dataframe containing original and treated column names of DF
}
\description{
cleanColNames : get colnames, remove leading and trailing whitespace and push to upper case
}

