% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_coerceData.R
\name{coerceData}
\alias{coerceData}
\title{coerceData}
\usage{
coerceData(doa, dob)
}
\arguments{
\item{doa}{Data object A (any object that can be coerced to a data frame)}

\item{dob}{Data object B (any object that can be coerced to a data frame)}
}
\value{
A list of 2 data frames, which is DOA and DOB coerced as data.frames
}
\description{
coerceData
}
\examples{
\dontrun{irisMatrix <- as.matrix(iris)}
\dontrun{coerceData(irisMatrix,iris)}
}
