% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_rCompObjectQueryFunctions.R
\name{colsWithUnequalValues}
\alias{colsWithUnequalValues}
\title{colsWithUnequalValues: a dataframe summarising a column with unequal values}
\usage{
colsWithUnequalValues(x, mismatches)
}
\arguments{
\item{x}{the column to be considered}

\item{mismatches}{- a mismatches object from an dataCompareR object}
}
\value{
data frame with a summary of the mismatching column
}
\description{
colsWithUnequalValues: a dataframe summarising a column with unequal values
}

