% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_matchColumns.R
\name{compareNames}
\alias{compareNames}
\title{compareNames : compare the intersect of colInfoA and colInfoB and return boolean of matched columns for each data frame}
\usage{
compareNames(colInfoA, colInfoB)
}
\arguments{
\item{colInfoA}{input data frames with original and treated column names}

\item{colInfoB}{input data frames with original and treated column names}
}
\description{
compareNames : compare the intersect of colInfoA and colInfoB and return boolean of matched columns for each data frame
}
