% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_matchRows.R
\name{createAntiSubset}
\alias{createAntiSubset}
\title{Create a dataframe of the rows that don't match}
\usage{
createAntiSubset(index_antisubset, original_keys, index_key, df)
}
\arguments{
\item{index_antisubset}{Vector of mismatching indices}

\item{original_keys}{A character array}

\item{index_key}{A character array}

\item{df}{A data frame}
}
\value{
A dataframe containing the dropped rows
}
\description{
Create a dataframe of the rows that don't match
}

