% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_createCleaningInfo.R
\name{createCleaningInfo}
\alias{createCleaningInfo}
\title{Converts cleaning info into a format consumable by updateCompareObject.}
\usage{
createCleaningInfo(compObj, cleaningInfo)
}
\arguments{
\item{compObj}{dataCompareRobject to be updated}

\item{cleaningInfo}{list of cleaning information}
}
\value{
\code{compObj} updated dataCompareRobject
}
\description{
Converts cleaning info into a format consumable by updateCompareObject.
}
