% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_createColMatching.R
\name{createColMatching}
\alias{createColMatching}
\title{Converts the output of the column matching logic to something consumable by updateCompareObject.}
\usage{
createColMatching(compObj, colMatchInfo)
}
\arguments{
\item{compObj}{dataCompareRobject instance to be updated}

\item{colMatchInfo}{List output from the column matching logic}
}
\value{
\code{compObj} updated with colMatching block
}
\description{
Converts the output of the column matching logic to something consumable by updateCompareObject.
}
