% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_createCompareObject.R
\name{createCompareObject}
\alias{createCompareObject}
\title{Generates an empty list of the correct class to store results}
\usage{
createCompareObject()
}
\value{
A list of class dataCompareRObject
}
\description{
Generates an empty list of the correct class to store results
}
