% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_createMeta.R
\name{createMeta}
\alias{createMeta}
\title{Takes the raw info for the meta block of the output
and puts it in a format usable by the updateCompareObject
function}
\usage{
createMeta(dataCompareRobject, DFA, DFB, arguments, timestamp)
}
\arguments{
\item{dataCompareRobject}{Object of class dataCompareRobject}

\item{DFA}{First data set passed in to the dataCompareR function}

\item{DFB}{Second data set passed in to the dataCompareR function}

\item{arguments}{Collection of arguments passed to compare object with labels that match the dataCompareR arg definitions}

\item{timestamp}{Timestamp}
}
\value{
\code{dataCompareRobject}
}
\description{
Takes the raw info for the meta block of the output
and puts it in a format usable by the updateCompareObject
function
}

