% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_createMismatches.R
\name{createMismatches}
\alias{createMismatches}
\title{Create mismatch object}
\usage{
createMismatches(compObj, misObj, keys)
}
\arguments{
\item{compObj}{RCompareObject, output from processFlow}

\item{misObj}{MismatchObject, output from compareData (processFlow)}

\item{keys}{Character vector, the keys matched on, to allow removal of any extra columns
introudced by the compare process}
}
\value{
The mismatch object
}
\description{
Create mismatch object
}

