% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_createMismatchObject.R
\name{createMismatchObject}
\alias{createMismatchObject}
\title{Create mismatch object}
\usage{
createMismatchObject(dat_a, dat_b, dat_eq, str_index)
}
\arguments{
\item{dat_a}{dataframe, output from prepareData}

\item{dat_b}{dataframe, output from prepareDate}

\item{dat_eq}{dataframe, output from locateMismatches}

\item{str_index, }{vector of index variables (could have length 1)}
}
\value{
An dataCompareR mismatch object
}
\description{
Create mismatch object
}
\examples{
\dontrun{createMismatchObject(dataA, dataB, mism, idx)}
}

