% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_createReportText.R
\name{createReportText}
\alias{createReportText}
\title{createReportText: prepares text which is used in the summary report
Saves R markdown and HTML reports in the area specified by the user. Reports are called RcompareReport.Rmd (.html)
Uses knitr package to create tables in the markdown (createReportText function) and HTML report.}
\usage{
createReportText(x)
}
\arguments{
\item{x}{input object which summary comparison information}
}
\value{
text in R markdown format
}
\description{
createReportText: prepares text which is used in the summary report
Saves R markdown and HTML reports in the area specified by the user. Reports are called RcompareReport.Rmd (.html)
Uses knitr package to create tables in the markdown (createReportText function) and HTML report.
}
\examples{
 
\dontrun{createReportText(x=MysummaryCompareObject)}
}

