% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_createRowMatching.R
\name{createRowMatching}
\alias{createRowMatching}
\title{function for updating a compare object with  
information passed to it from the match rows function}
\usage{
createRowMatching(compObj, x, matchKey)
}
\arguments{
\item{compObj}{dataCompareRobject to be updated}

\item{x}{Object of information with classes related to the relevant section of the dataCompareRobject}

\item{matchKey}{the list of keys based on which the row matching was performed}
}
\value{
compObj Updated dataCompareRobject
}
\description{
function for updating a compare object with  
information passed to it from the match rows function
}

