% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_unifiedOutputGenerator.R
\name{createTextSummary}
\alias{createTextSummary}
\title{createTextSummary: create a text based summary of an dataCompareR object}
\usage{
createTextSummary(x, ...)
}
\arguments{
\item{x}{an dataCompareR object}

\item{...}{Arguments passed on to other functions}
}
\value{
cat's lines to the screen (or to be captured)
 cat(newLine)
}
\description{
createTextSummary: create a text based summary of an dataCompareR object
}

