% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_currentObjVersion.R
\name{currentObjVersion}
\alias{currentObjVersion}
\title{Place to store and access the current object version.}
\usage{
currentObjVersion()
}
\value{
\code{currentVersion} int of the version number
}
\description{
Place to store and access the current object version.
}
