% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_executeCoercions.R
\name{executeCoercions}
\alias{executeCoercions}
\title{executeCoercions:}
\usage{
executeCoercions(DFA, DFB, WhitespaceTrim = T)
}
\arguments{
\item{DFA}{Input dataframe A}

\item{DFB}{Input dataframe B}

\item{WhitespaceTrim}{User defined boolean for whether leading/trailing white space is trimmed in strings (TRUE / FALSE)}
}
\value{
\code{out} list containing 3 data frames DFA, DFB and DataTypes

\code{DFA} Dataframe with factor fields converted to character type and white space trimming (if option is selected by the user)

\code{DFB} Dataframe with factor fields converted to character type and white space trimming (if option is selected by the user)

\code{DataTypes} Dataframe with field types before and after cleaning for both DFA and DFB
}
\description{
executeCoercions:
}
\examples{
\dontrun{executeCoercions(DFA=iris,DFB=iris,WhitespaceTrim=T)}
}
