% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_generateMismatchData.R
\name{generateMismatchData}
\alias{generateMismatchData}
\title{Extract data from a dataCompareR comparison}
\usage{
generateMismatchData(x, dfA, dfB, ...)
}
\arguments{
\item{x}{A dataCompareRobject.}

\item{dfA}{Data frame (or object coercable to a data frame). One of the two data frames used in the initial rCompare call.}

\item{dfB}{Data frame (or object coercable to a data frame). One of the two data frames used in the initial rCompare call.}

\item{...}{Unused currently, may be used in future}
}
\value{
\code{mismatchData} A list containing two objects: mismatched rows in first data object and mismatched rows in 
second data object
}
\description{
Produces a list of two data frames, containing the mismatched rows from the two input tables

Note that this function requires the user to pass in the two data frames used in the inital comparison. If this data
does not match that used for the generation of the dataCompareR object the results produced will not be accurate.
}
\seealso{
Other dataCompareR.functions: \code{\link{print.dataCompareRobject}},
  \code{\link{rCompare}}, \code{\link{saveReport}},
  \code{\link{summary.dataCompareRobject}}
}
\concept{dataCompareR.functions}
