% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_createCleaningInfo.R
\name{getCoercions}
\alias{getCoercions}
\title{Subsets on the variables that have a coercion.}
\usage{
getCoercions(typesDf)
}
\arguments{
\item{typesDf}{Dataframe of type information from the executeCoercion function}
}
\value{
\code{coercedT} Subset version of typesDf where a coercion occurred
}
\description{
Subsets on the variables that have a coercion.
}
