% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_createColMatching.R
\name{getMismatchColNames}
\alias{getMismatchColNames}
\title{Extracts the column names only in one data frame from a table of 
match information}
\usage{
getMismatchColNames(colMatchInfo, colNameCol, matchFlagCol)
}
\arguments{
\item{colMatchInfo}{Dataframe with column names, match flag}

\item{colNameCol}{Name of the column with the column names}

\item{matchFlagCol}{Name of the column with the match flag}
}
\value{
Vector of column names that do not match
}
\description{
Extracts the column names only in one data frame from a table of 
match information
}

