% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_isRCompareObject.R
\name{is.dataCompareRobject}
\alias{is.dataCompareRobject}
\title{Check object is of class dataCompareRobject}
\usage{
is.dataCompareRobject(x)
}
\arguments{
\item{x}{An object}
}
\value{
A boolean: TRUE if object is class dataCompareRobject and FALSE if not
}
\description{
Check object is of class dataCompareRobject
}

