% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_helperFunctions.R
\name{isNotNull}
\alias{isNotNull}
\title{isNotNull: is object not null}
\usage{
isNotNull(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{boolean} is object null T/F
}
\description{
isNotNull: is object not null
}
\examples{
\dontrun{isNotNull(NULL)}
\dontrun{isNotNull(5)}
}
