% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_checkEmpty.R
\name{isSingleNA}
\alias{isSingleNA}
\title{isSingleNA}
\usage{
isSingleNA(x)
}
\arguments{
\item{x}{literally anything}
}
\value{
boolean
}
\description{
Boolean function - T if x is a single NA. False otherwise.
}
