% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_outputStructureFunctions.R
\name{listObsNotVerbose}
\alias{listObsNotVerbose}
\title{listObsNotVerbose}
\usage{
listObsNotVerbose(i, x, uniquevarlist, nObs)
}
\arguments{
\item{i}{The position of the element we want to compare}

\item{x}{An dataCompareR object}

\item{uniquevarlist}{A list of the variables in the compare}

\item{nObs}{How many obervations to return}
}
\value{
A list of mismatching observations from start/end of mismatches
}
\description{
Return a summary of mismatching data
}

