% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_outputStructureFunctions.R
\name{listObsVerbose}
\alias{listObsVerbose}
\title{listObsVerbose}
\usage{
listObsVerbose(i, x)
}
\arguments{
\item{i}{The position of the element we want to compare}

\item{x}{An dataCompareR object}
}
\value{
A list of mismatching observations
}
\description{
Return all mismatching data
}
