% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_locateMismatches.R
\name{locateMismatches}
\alias{locateMismatches}
\title{Checks whether elements in two input data frames are equal.}
\usage{
locateMismatches(DFA, DFB, keys = NULL, maxMismatches = NA)
}
\arguments{
\item{DFA}{input data frame}

\item{DFB}{input data frame}

\item{keys}{character vector of index variables}

\item{maxMismatches}{Integer. The max number of mismatches to assess, after which dataCompareR will stop 
(without produceing an dataCompareR object). Designed to improve performance for large datasets.}
}
\value{
data frame containing keys and boolean logic of match/no match for each element
        If data types are not equal returns FALSE. Treats NA and NaN as unequal.
}
\description{
Checks whether elements in two input data frames are equal.
}

