% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_validateArguments.R
\name{makeValidKeys}
\alias{makeValidKeys}
\title{makeValidKeys}
\usage{
makeValidKeys(keys)
}
\arguments{
\item{keys}{A character vector}
}
\value{
A character vector with syntactically valid names
}
\description{
Correct syntactically invalid Keys
}
\examples{
\dontrun{makeValidKeys(c(" hello", "__BAD NAME___")}
}
