% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_validateArguments.R
\name{makeValidNames}
\alias{makeValidNames}
\title{makeValidNames}
\usage{
makeValidNames(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame with syntactically valid names
}
\description{
Correct syntactically invalid names in a data frame
}
\examples{
\dontrun{makeValidNames(iris)}
}
