% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_matchColumns.R
\name{matchColumns}
\alias{matchColumns}
\title{matchColumns : create subset of DFA and DFB to contain matching column names for both data frames}
\usage{
matchColumns(DFA, DFB)
}
\arguments{
\item{DFA}{input data frame}

\item{DFB}{input data frame}
}
\value{
matchColOut named list of data frames. subsetA,subsetB contain only columns common to both data frames. colInfoA,colInfoB contain mapping of column names from original to treated and boolean indicator of common columns.
}
\description{
matchColumns : create subset of DFA and DFB to contain matching column names for both data frames
}
