% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_matchRows.R
\name{matchMultiIndex}
\alias{matchMultiIndex}
\title{Generate two dataframes that contain the same rows based on a two-column index}
\usage{
matchMultiIndex(df_a, df_b, indices)
}
\arguments{
\item{df_a}{A dataframe}

\item{df_b}{A dataframe}

\item{indices}{A char vector}
}
\value{
A list containing the two dataframes, subsetted by shared indices, and a list which itself 
       contains the vectors for the dropped rows
}
\description{
Generate two dataframes that contain the same rows based on a two-column index
}

