% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_matchRows.R
\name{matchNoIndex}
\alias{matchNoIndex}
\title{Generate two dataframes that contain the same rows based on a two-column index}
\usage{
matchNoIndex(df_a, df_b)
}
\arguments{
\item{df_a}{A dataframe}

\item{df_b}{A dataframe}
}
\value{
A list containing the two dataframes, subsetted to the size of the smaller one, and
         a list containing vectors of the rows dropped.
}
\description{
Generate two dataframes that contain the same rows based on a two-column index
}
