% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_matchRows.R
\name{matchRows}
\alias{matchRows}
\title{Generate two dataframes and returns subsets of these dataframes that have shared rows.}
\usage{
matchRows(df_a, df_b, indices = NA)
}
\arguments{
\item{df_a}{A dataframe}

\item{df_b}{A dataframe}

\item{indices}{The indices to match rows between \code{df_a} and \code{df_b}. Can be NA,
single character, or a vector of characters}
}
\value{
A list containing the two dataframes, subsetted by shared indices, and a list which itself 
       contains dataframes for the dropped rows
}
\description{
Generate two dataframes and returns subsets of these dataframes that have shared rows.
}

