% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_matchRows.R
\name{matchSingleIndex}
\alias{matchSingleIndex}
\title{Generate two dataframes that contain the same rows based on a single index}
\usage{
matchSingleIndex(df_a, df_b, index_key, original_keys)
}
\arguments{
\item{df_a}{A dataframe}

\item{df_b}{A dataframe}

\item{index_key}{A character vector}

\item{original_keys}{A character vector}
}
\value{
A list containing the two dataframes, subsetted by shared indices, and a list which itself
       contains the vectors for the dropped rows
}
\description{
Generate two dataframes that contain the same rows based on a single index
}
