% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_locateMismatches.R
\name{mismatchHighStop}
\alias{mismatchHighStop}
\title{mismatchHighStop Checks if we've exceeded threshold of mismatches}
\usage{
mismatchHighStop(trueFalseMatrix, maxMismatches)
}
\arguments{
\item{trueFalseMatrix}{a matrix of true/false}

\item{maxMismatches}{number of mismatches at which the routine stops}
}
\value{
Nothing. Stops if threshold exceeded
}
\description{
mismatchHighStop Checks if we've exceeded threshold of mismatches
}
