% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_matchColumns.R
\name{orderColumns}
\alias{orderColumns}
\title{orderColumns: order columns by treated column names}
\usage{
orderColumns(colInfo)
}
\arguments{
\item{colInfo}{dataframe containing original and treated column names of DF}
}
\value{
ordered colInfo dataframe containing original and treated column names of DF
}
\description{
orderColumns: order columns by treated column names
}

