% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_helperFunctions.R
\name{outputSectionHeader}
\alias{outputSectionHeader}
\title{outputSectionHeader: creates an outputSectionHeader}
\usage{
outputSectionHeader(headerName)
}
\arguments{
\item{headerName}{a header name}
}
\value{
\code{character} a character based section headers
}
\description{
outputSectionHeader: creates an outputSectionHeader
}
