% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_printsummaryrcompareobject.R
\name{print.summary.dataCompareRobject}
\alias{print.summary.dataCompareRobject}
\title{Printing summaryRCompare Output}
\usage{
\method{print}{summary.dataCompareRobject}(x, ...)
}
\arguments{
\item{x}{an object of class "summary.dataCompareRobject", usually a result of a call to 
\code{\link{summary.dataCompareRobject}}.}

\item{...}{Additional arguments passsed on to \code{\link{createTextSummary}}}
}
\description{
Printing summaryRCompare Output
}
\examples{
rc1 <- rCompare(iris,iris)
summary(rc1)  
}

