% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_rounddf.R
\name{rounddf}
\alias{rounddf}
\title{Round all numeric fields in a data frame}
\usage{
rounddf(df, roundDigits)
}
\arguments{
\item{df}{A data frame to round}

\item{roundDigits}{Number of digits to round to}
}
\value{
A rounded data frame
}
\description{
Round all numeric fields in a data frame
}
