% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_updateCompareObject_CleaningInfo.R
\name{updateCompareObject.cleaninginfo}
\alias{updateCompareObject.cleaninginfo}
\title{Updates cleaning info in the compare object}
\usage{
\method{updateCompareObject}{cleaninginfo}(x, compObj)
}
\arguments{
\item{x}{list of type cleaninginfo with data types}

\item{compObj}{dataCompareRobject to be updated}
}
\value{
\code{compObj} updated dataCompareRobject
}
\description{
Updates cleaning info in the compare object
}
