% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_updateCompareObject_ColMatching.R
\name{updateCompareObject.colmatching}
\alias{updateCompareObject.colmatching}
\title{Adds a colMatching block to the output}
\usage{
\method{updateCompareObject}{colmatching}(x, compObj)
}
\arguments{
\item{x}{List of class colmatching with column matching info}

\item{compObj}{dataCompareRobject instance to be updated}
}
\value{
\code{compObj} Updated dataCompareRobject
}
\description{
Adds a colMatching block to the output
}
