% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_updateCompareObject_Matches.R
\name{updateCompareObject.matches}
\alias{updateCompareObject.matches}
\title{Adds a colMatching block to the output}
\usage{
\method{updateCompareObject}{matches}(x, compObj)
}
\arguments{
\item{x}{List of class 'matches' with column matching info}

\item{compObj}{dataCompareRobject instance to be updated}
}
\value{
\code{compObj} Updated dataCompareRobject
}
\description{
Adds a colMatching block to the output
}
