% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_updateCompareObject_Meta.R
\name{updateCompareObject.meta}
\alias{updateCompareObject.meta}
\title{Takes raw info for meta and adds it to the compare object}
\usage{
\method{updateCompareObject}{meta}(x, compObj)
}
\arguments{
\item{x}{List of class 'meta' with data related to meta}

\item{compObj}{dataCompareRobject to be appended}
}
\value{
\code{compObj} dataCompareRobject updated with meta block
}
\description{
Takes raw info for meta and adds it to the compare object
}

