% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_updateCompareObject_Mismatches.R
\name{updateCompareObject.mismatches}
\alias{updateCompareObject.mismatches}
\title{Adds a colMatching block to the output}
\usage{
\method{updateCompareObject}{mismatches}(x, compObj)
}
\arguments{
\item{x}{List of class 'mismatches' with column matching info}

\item{compObj}{dataCompareRobject instance to be updated}
}
\value{
\code{compObj} Updated dataCompareRobject
}
\description{
Adds a colMatching block to the output
}
