% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_updateCompareObject.R
\name{updateCompareObject}
\alias{updateCompareObject}
\title{Generic function for updating a compare object with 
information passed to it, that has methods based on the class
of the info argument.}
\usage{
updateCompareObject(x, compObj)
}
\arguments{
\item{x}{Object of information with classes related to the relevant section of the dataCompareRobject}

\item{compObj}{dataCompareRobject to be updated}
}
\value{
compObj Updated dataCompareRobject
}
\description{
Generic function for updating a compare object with 
information passed to it, that has methods based on the class
of the info argument.
}

