% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_updateCompareObject_RowMatching.R
\name{updateCompareObject.rowmatching}
\alias{updateCompareObject.rowmatching}
\title{Adds a rowMatching block to the output}
\usage{
\method{updateCompareObject}{rowmatching}(x, compObj)
}
\arguments{
\item{x}{List of class rowMatching with row matching info}

\item{compObj}{dataCompareRobject instance to be updated}
}
\value{
\code{compObj} Updated dataCompareRobject
}
\description{
Adds a rowMatching block to the output
}
