% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_validateArguments.R
\name{validateArguments}
\alias{validateArguments}
\title{validateArguments}
\usage{
validateArguments(matchKey = NA, roundDigits = NA, coerceCols = TRUE,
  maxMismatch = NA)
}
\arguments{
\item{matchKey}{A character or character vector of column names to match on}

\item{roundDigits}{Integer. If NA, numerics are not rounded before comparison. If /code{roundDigits} is specified, numerics are 
rounded to /code{roundDigits} decimal places using \link[base]{round}.}

\item{coerceCols}{Boolean - do we coerce columns names?}

\item{maxMismatch}{Cap for number of mismatches}
}
\value{
Nothing. Errors if any parameters are invalid.
}
\description{
validateArguments
}
\examples{
\dontrun{validateArguments('plantName',1E-8,T,1000)}
\dontrun{validateArguments('colorName',1E-9,F,10)}
}

