% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_validateData.R
\name{validateData}
\alias{validateData}
\title{validateData : routine to validate the input data}
\usage{
validateData(df1, df2, keys = NA)
}
\arguments{
\item{df1}{a data frame}

\item{df2}{a data frame}

\item{keys}{Keys used}
}
\value{
None. Stops if error.
}
\description{
validateData : routine to validate the input data
}
\examples{
\dontrun{validateData(iris,iris)}
}
