% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_createMismatchObject.R
\name{variableDetails}
\alias{variableDetails}
\title{Create variable mismatch details}
\usage{
variableDetails(dat)
}
\arguments{
\item{dat}{The mismatch data}
}
\value{
mismatch details
}
\description{
Create variable mismatch details
}
