% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_createMismatchObject.R
\name{variableMismatches}
\alias{variableMismatches}
\title{Create variable mismatch table}
\usage{
variableMismatches(varname, vals_a, vals_b, vector_eq)
}
\arguments{
\item{varname, }{variable to create mismatch table for}

\item{vals_a, }{variables from dfA}

\item{vals_b, }{variables from dfB}

\item{vector_eq, }{a list of columns which are equal}
}
\value{
Mismatch table
}
\description{
Create variable mismatch table
}

