% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_rCompare.R
\name{warnLargeData}
\alias{warnLargeData}
\title{Warn users if the calculation is likely to be slow}
\usage{
warnLargeData(nrow_dfa, ncol_dfa, nrow_dfb, ncol_dfb)
}
\arguments{
\item{nrow_dfa}{number of rows in first data frame}

\item{ncol_dfa}{number of columns in first data frame}

\item{nrow_dfb}{number of rows in second data frame}

\item{ncol_dfb}{number of columns in second data frame}
}
\value{
Nothing
}
\description{
Checks if there are more than 20E6 elements for comparison. If there are, spits out a warning 
message that the calculation may run slowly
}
