% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_rCompObjectQueryFunctions.R
\name{rcompObjItemLength}
\alias{rcompObjItemLength}
\title{rcompObjItemLength: return length of an item, returning 0 if null}
\usage{
rcompObjItemLength(x)
}
\arguments{
\item{x}{an object}
}
\value{
length, numeric
}
\description{
rcompObjItemLength: return length of an item, returning 0 if null
}

